+++ ElvUI_MerathilisUI/Core/Changelog/7.04.lua
@@ -1,8 +1,12 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[704] = {
-	RELEASE_DATE = "TBD",
+	RELEASE_DATE = "17.01.2026",
 	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	NEW = {
+		"[Scale]: Added option to scale the Encounter Journal",
+	},
+	IMPROVEMENTS = {
+		"[Vehicle Bar]: Bring back the Vigor Bar and Speed Text",
+	},
 }